refactor: Incorporating renauter feedback / better modules partitioning
