+++ substrate-node/pallets/pallet-dao/src/benchmarking.rs
@@ -1,7 +1,8 @@
 #![cfg(feature = "runtime-benchmarks")]
 
-use super::*;
-use crate::Pallet as DaoModule;
+use crate::pallet::Pallet as DaoModule;
+use crate::pallet::Pallet;
+use crate::pallet::{Call, Config, Event};
 use frame_benchmarking::{account, benchmarks, whitelisted_caller};
 use frame_support::{assert_ok, BoundedVec};
 use frame_system::{Call as SystemCall, EventRecord, Pallet as System, RawOrigin};
