+++ substrate-node/pallets/pallet-dao/src/proposal.rs
@@ -2,6 +2,8 @@ use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
 use sp_std::vec::Vec;
 
+pub type ProposalIndex = u32;
+
 #[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Debug, Encode, Decode, Default, TypeInfo)]
 pub struct DaoProposal<ProposalIndex> {
     pub index: ProposalIndex,
